requirements.txt
asgiref==3.2.10
coverage==5.3
Django==3.1.1
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
bcrypt~=3.2.0
django-author~=1.0.2
djangorestframework~=3.12.1